+++ common/src/main/java/com/avp/common/entity/AVPEntityTypeTags.java
@@ -16,6 +16,8 @@ public record AVPEntityTypeTags() {
 
     public static final TagKey<EntityType<?>> ANIMALS = create("animals");
 
+    public static final TagKey<EntityType<?>> CHESTBURSTERS = create("chestbursters");
+
     public static final TagKey<EntityType<?>> DRONES = create("drones");
 
     public static final TagKey<EntityType<?>> FACEHUGGERS = create("facehuggers");